Add explicit handling for removing all items from the cache (via CLI). Move main CLI definition into the cli package.
